# backend/openapi.yaml
openapi: 3.0.0
info:
  title: PropertyPilot Backend API
  description: Real estate investment analysis platform with financial calculations, property management, and reporting.
  version: 1.0.0
  contact:
    name: PropertyPilot Support
    email: support@propertypilot.com
  license:
    name: MIT
servers:
  - url: https://api.propertypilot.com
    description: Production server
  - url: http://localhost:3001
    description: Development server
...
paths:
...
  /api/properties:
    get:
      tags:
        - Properties
      summary: List all properties
      description: Get all properties for the authenticated user with filtering and pagination
      security:
        - SessionAuth: []
      parameters:
        - name: type
          in: query
          description: Filter by property type
          schema:
            type: string
            enum: [rent, flip, airbnb]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [lead, analyzing, offer, under_contract, owned, sold]
        - name: city
          in: query
          description: Filter by city (partial match)
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of properties
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Property'
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  hasMore:
                    type: boolean
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Properties
      summary: Create property
      description: Create a new property
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyRequest'
      responses:
        '201':
          description: Property created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      property:
                        $ref: '#/components/schemas/Property'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/properties/{propertyId}:
    get:
      tags:
        - Properties
      summary: Get property detail
      description: Get detailed information about a specific property
      security:
        - SessionAuth: []
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Property found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      property:
                        $ref: '#/components/schemas/Property'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - Properties
      summary: Update property
      description: Update property details
      security:
        - SessionAuth: []
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PropertyUpdateRequest'
      responses:
        '200':
          description: Property updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      property:
                        $ref: '#/components/schemas/Property'

    delete:
      tags:
        - Properties
      summary: Delete property
      description: Delete a property (soft delete)
      security:
        - SessionAuth: []
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Property deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Property deleted successfully
...
  /api/properties/{propertyId}/renovations:
    post:
      tags:
        - Renovations
      summary: Create renovation item
      description: Add a renovation item to a property
      security:
        - SessionAuth: []
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenovationItemRequest'
      responses:
        '201':
          description: Renovation created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      renovation:
                        $ref: '#/components/schemas/RenovationItem'
  /api/renovations/{renovationId}:
    put:
      tags:
        - Renovations
      summary: Update renovation item
      description: Update a renovation item
      security:
        - SessionAuth: []
      parameters:
        - name: renovationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenovationItemRequest'
      responses:
        '200':
          description: Renovation updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      renovation:
                        $ref: '#/components/schemas/RenovationItem'

    delete:
      tags:
        - Renovations
      summary: Delete renovation item
      description: Delete a renovation item
      security:
        - SessionAuth: []
      parameters:
        - name: renovationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Renovation deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Renovation deleted successfully
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: sessionId
      description: Session-based authentication via httpOnly cookie
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        address:
          type: string
        city:
          type: string
        state:
          type: string
        zipCode:
          type: string
        type:
          type: string
          enum: [rent, flip, airbnb]
        status:
          type: string
          enum: [lead, analyzing, offer, under_contract, owned, sold]
        purchasePrice:
          type: number
          format: float
        currentValue:
          type: number
          format: float
        arv:
          type: number
          format: float
        sqft:
          type: number
          format: float
        bedrooms:
          type: integer
        bathrooms:
          type: number
          format: float
        yearBuilt:
          type: integer
        lotSize:
          type: number
          format: float
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PropertyRequest:
      type: object
      properties:
        address:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 2
        zipCode:
          type: string
          minLength: 5
        type:
          type: string
          enum: [rent, flip, airbnb]
        status:
          type: string
          enum: [lead, analyzing, offer, under_contract, owned, sold]
        purchasePrice:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        currentValue:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        arv:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        sqft:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        bedrooms:
          type: integer
          minimum: 1
        bathrooms:
          type: number
          format: float
          minimum: 1
        yearBuilt:
          type: integer
        lotSize:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        notes:
          type: string
      required:
        - address
        - city
        - state
        - zipCode
        - type
        - status
        - purchasePrice
        - currentValue
        - arv
        - sqft
        - bedrooms
        - bathrooms
        - yearBuilt
        - lotSize
        - latitude
        - longitude
    PropertyUpdateRequest:
      type: object
      properties:
        address:
          type: string
          minLength: 1
        city:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 2
        zipCode:
          type: string
          minLength: 5
        type:
          type: string
          enum: [rent, flip, airbnb]
        status:
          type: string
          enum: [lead, analyzing, offer, under_contract, owned, sold]
        purchasePrice:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        currentValue:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        arv:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        sqft:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        bedrooms:
          type: integer
          minimum: 1
        bathrooms:
          type: number
          format: float
          minimum: 1
        yearBuilt:
          type: integer
        lotSize:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        notes:
          type: string
...
    RenovationItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        propertyId:
          type: string
          format: uuid
        category:
          type: string
        description:
          type: string
        estimatedCost:
          type: number
          format: float
        actualCost:
          type: number
          format: float
          nullable: true
        contractor:
          type: string
          nullable: true
        startDate:
          type: string
          format: date-time
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RenovationItemRequest:
      type: object
      properties:
        category:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        estimatedCost:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        actualCost:
          type: number
          format: float
          minimum: 0
          exclusiveMinimum: true
        contractor:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, in_progress, completed]
        notes:
          type: string
      required:
        - category
        - description
        - estimatedCost
        - status
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Validation failed
              details:
                type: array
                items:
                  type: object
                  properties:
                    path:
                      type: array
                      items:
                        type: string
                    message:
                      type: string
              requestId:
                type: string
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Unauthorized
              requestId:
                type: string
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Resource not found
              requestId:
                type: string